HC_API_Plat/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.2
Flask-JWT-Extended==4.7.1
python-dotenv==1.0.0
psycopg2-binary==2.9.6
Werkzeug>=2.3.7
pybars3